added ext MOX, & encoder increment settable
Ext MOX input enabled.
Encoders can now be programmes for the number of edges/events per "detent" position; this generally depends on the encoder used. By default should be set to 2 for the dual encoders, and we recommend starting with 4 for the VFO encoder.
The settings are on the console config page.